% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerage.R
\name{player_oldest}
\alias{player_oldest}
\title{An Oldest Player Function}
\usage{
player_oldest(year)
}
\arguments{
\item{year}{Year of interest.}
}
\description{
This function allows you to see which player was the oldest in a specific year
}
\examples{
player_oldest(2005)
}
\keyword{age}
